fixup! [#741] separation ofbase eventAPI  to EventDao and eventApi
