+++ src/features/Events/EventApi.ts
@@ -1,12 +1,8 @@
-import { extractCalendarHrefs } from '@/utils/extractCalendarHrefs'
 import { extractEventBaseUuid } from '@/utils/extractEventBaseUuid'
-import { isValidEmail } from '@/utils/isValidEmail'
 import { convertEventDateTimeToISO, resolveTimezoneId } from '@/utils/timezone'
 import { TIMEZONES } from '@/utils/timezone-data'
 import ICAL from 'ical.js'
-import { hasFreeBusyConflict } from '../../components/Attendees/useFreeBusy'
 import { CalDavItem } from '../Calendars/api/types'
-import { getCalendars } from '../Calendars/CalendarApi'
 import { Calendar } from '../Calendars/CalendarTypes'
 import {
   VCalComponent,
@@ -20,10 +16,6 @@ import {
   fetchAllRecurrentVevents,
   fetchEventIcs,
   fetchEventRaw,
-  fetchFreeBusyPost,
-  fetchFreeBusyReports,
-  fetchUserByEmail,
-  FreeBusyPostQuery,
   importEventRaw,
   moveEventRaw,
   postCounterProposalRaw,
@@ -374,70 +366,6 @@ export async function searchEvent(
     reqParam.attendees = attendees
   }
   return searchEventRaw(reqParam)
-}
-
-export async function getFreeBusyForAddedAttendees(
-  userId: string,
-  start: string,
-  end: string
-): Promise<boolean> {
-  const calendars = await getCalendars(
-    userId,
-    'withFreeBusy=true&withRights=true'
-  )
-  const hrefs = extractCalendarHrefs(calendars)
-  if (hrefs.length === 0) return false
-
-  const results = await fetchFreeBusyReports({ hrefs, start, end })
-  return results.some(data => (data ? hasFreeBusyConflict(data) : false))
-}
-
-interface BusySlot {
-  uid: string
-  start: string
-  end: string
-}
-
-interface CalendarFreeBusy {
-  id: string
-  busy: BusySlot[]
-}
-
-interface UserFreeBusy {
-  id: string
-  calendars: CalendarFreeBusy[]
-}
-interface FreeBusyPayload {
-  users?: UserFreeBusy[]
-}
-
-export async function getFreeBusyForEventAttendees(
-  userIds: string[],
-  start: string,
-  end: string,
-  eventUid: string
-): Promise<Record<string, boolean>> {
-  const query: FreeBusyPostQuery = { userIds, start, end, eventUid }
-  const payload = (await fetchFreeBusyPost(query)) as FreeBusyPayload
-  const users = Array.isArray(payload.users) ? payload.users : []
-  const eventUidBase = extractEventBaseUuid(eventUid)
-  return Object.fromEntries(
-    users.map(u => {
-      const isBusy = (u.calendars ?? []).some(cal =>
-        (cal.busy ?? []).some(
-          slot => extractEventBaseUuid(slot.uid) !== eventUidBase
-        )
-      )
-      return [u.id, isBusy]
-    })
-  )
-}
-
-export async function getUserDataFromEmail(
-  email: string
-): Promise<Array<Record<string, string>>> {
-  if (!isValidEmail(email)) return []
-  return fetchUserByEmail(email)
 }
 
 export async function postCounterProposal({
